Moved firm properties into a separate file.
