+++ .gitignore
@@ -36,3 +36,4 @@ model/__pycache__/__init__.cpython-311.pyc
 model/economy/trade/__pycache__/__init__.cpython-311.pyc
 model/economy/trade/__pycache__/market.cpython-311.pyc
 model/population/__pycache__/group_properties.cpython-311.pyc
+model/economy/industry/__pycache__/firm_properties.cpython-311.pyc
