+++ model/economy/industry/firm.py
@@ -1,25 +1,11 @@
 from model.core.random import _sample_normal
 from dataclasses import dataclass
 
-@dataclass
-class FirmParams:
-    productivity: float
-    production_capacity: int
-    ownership: str
-    good: str
-    # transitional: optional in input_data.json
-    capital: float | None = None
-    wage: float | None = None
-    country_policy: dict | None = None
+from model.economy.industry.firm_properties import (FirmParams,
+                                                    FirmState,
+                                                    FirmProperties)
 
-@dataclass
-class FirmState:
-    employed: int = 0
-    total_productivity: float = 0.0
-    inventory: float = 0.0
-    market_capital: float = 0.0
-
-class Firm:
+class Firm(FirmProperties):
     def __init__(self, params: FirmParams, rng):
         self.p = params
         self.state = FirmState()
@@ -42,43 +28,6 @@ class Firm:
             ),
             rng=rng,
         )
-
-
-    @property
-    def employed(self):
-        return self.state.employed
-    @employed.setter
-    def employed(self, value):
-        self.state.employed = value
-
-    @property
-    def total_productivity(self):
-        return self.state.total_productivity
-    @total_productivity.setter
-    def total_productivity(self, value):
-        self.state.total_productivity = value
-
-    @property
-    def inventory(self):
-        return self.state.inventory
-    @inventory.setter
-    def inventory(self, value):
-        self.state.inventory = value
-
-    @property
-    def market_capital(self):
-        return self.state.market_capital
-    @market_capital.setter
-    def market_capital(self, value):
-        self.state.market_capital = value
-
-    @property
-    def good(self):
-        return self.p.good
-
-    @property
-    def ownership(self):
-        return self.p.ownership
 
     def labour_demand(self, market_capital: float | None = None, market_wage: float | None = None):
         '''Limiting factor of employment is either:
@@ -103,9 +52,6 @@ class Firm:
 
     def produce(self):
         self.inventory += _sample_normal(expected=self.update_total_productivity(), rng=self.rng)
-
-        self.market_capital -= self.employed * self.p.wage
-
 
     def tick(self):
         self.produce()
